requirements.txt
Django==3.2.12
djangorestframework==3.13.1
mysqlclient==2.1.0
django-mysql==4.5.0
django-cors-headers==3.11.0
pika==1.2.0